+++ game/HOWL/h50_SongPool_CalculateTempo.c
@@ -1,6 +1,7 @@
 #include <common.h>
 
-int DECOMP_SongPool_CalculateTempo(int const60, int tpqn, int bpm)
+// NOTE(aalhendi): ASM-verified NTSC-U 926 0x8002a678-0x8002a6cc
+u32 DECOMP_SongPool_CalculateTempo(s16 const60, s16 tpqn, s16 bpm)
 {
 	return (((tpqn * bpm) / 60) << 0x10) / const60;
 }
